refactor: best11 api 파라미터 타입 List<Long>으로 변경#191
